rename cohere colors export
