+++ src/interfaces/assistants_web/.eslintrc.js
@@ -1,5 +1,9 @@
 module.exports = {
-  extends: ['eslint-config-next/core-web-vitals', 'eslint-config-prettier', 'plugin:storybook/recommended'],
+  extends: [
+    'eslint-config-next/core-web-vitals',
+    'eslint-config-prettier',
+    'plugin:storybook/recommended',
+  ],
   rules: {
     '@typescript-eslint/no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],
     'no-process-env': 'error',
