+++ src/interfaces/assistants_web/src/utils/cohereColors.ts
